@@ -8,8 +8,8 @@
 	<packaging>jar</packaging>
 
 	<properties>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>16</maven.compiler.source>
+		<maven.compiler.target>16</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<geotools.version>23.5</geotools.version>
 	</properties>
@@ -48,9 +48,9 @@
 				<version>2.4</version>
 				<configuration>
 					<compilerVersion>1.8</compilerVersion>
-					<source>11</source>
+					<source>16</source>
 <!--					<complianceLevel>1.8</complianceLevel>-->
-					<target>11</target>
+					<target>16</target>
 					<includes>
 						<include>**/*</include>
 					</includes>
@@ -160,7 +160,7 @@
 		<dependency>
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-lang3</artifactId>
-			<version>3.10</version>
+			<version>3.12.0</version>
 		</dependency>
 		<dependency>
 			<groupId>com.opencsv</groupId>
@@ -170,7 +170,7 @@
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-databind</artifactId>
-			<version>2.10.5.1</version>
+			<version>2.11.1</version>
 		</dependency>
 	</dependencies>
 	<repositories>
